% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/send_gmail.R
\name{send_gmail}
\alias{send_gmail}
\title{Send an email}
\usage{
send_gmail(subject = "", body = "", to=NULL, body_sep = "\\\\n",
  file_private = ".gmail_private", dir_private = NULL)
}
\arguments{
\item{subject}{Character string with subject for the email
(Can't include single-quotes.)}

\item{body}{Character string or vector of character strings, with
    the body of the email; if it's a vector, it's combined with
    \code{collapse=body_sep} (default = \code{"\\n"}).
    (Contents can't include single-quotes, and you need to
    double-backslashed the backslashes.)}

\item{to}{Character string with email address of recipient; if
NULL, send to self (taken from private info)}

\item{body_sep}{Character string for combining elements in the input \code{body}.}

\item{file_private}{Character string with file name containing the private information.}

\item{dir_private}{The directory containing \code{file}. If NULL,
    we use the user's home directory \code{"~"}. Use \code{""}
    if the \code{file} argument includes the directory.}
}
\value{
Returns value from \code{\link[gmailR]{gmail}}.
}
\description{
Send an email, using google mail info from a private file
}
\details{
Uses gmail address and password information in the
\code{~/.gmail_private} file (or another file, specified with
\code{dir_private} and \code{file_private}. This should contain records
\code{gmail}, \code{password}, and \code{text}, and (optionally) \code{to}.

Email will be sent from the address defined in the private file as \code{gmail}.
If to is defined in that file, it is used as the default "to" email address;
otherwise, \code{gmail} is used as both the "from" and the "to" addresses.
You can specify a different recipient with the `to` argument to the function.
}
\examples{
\dontrun{
send_gmail("R message", "R process done")
send_gmail("R message", "R process done", to="someone@somewhere.com")
}
}
\seealso{
\code{\link{send_text}}
}
\keyword{utilities}
