% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/send_text.R
\name{send_text}
\alias{send_text}
\title{Send a text to yourself}
\usage{
send_text(subject = "", body = "", body_sep = "\\\\n",
  file_private = ".gmail_private", dir_private = NULL)
}
\arguments{
\item{subject}{Character string with subject for the email
(Can't include single-quotes.)}

\item{body}{Character string or vector of character strings, with
the body of the email; if it's a vector, it's combined with
\code{collapse=body_sep} (default = \code{"\\n"}).
(Contents can't include single-quotes, and you need to
double-backslashed the backslashes.)}

\item{body_sep}{Character string for combining elements in the input \code{body}.}

\item{file_private}{Character string with file name containing the private information.}

\item{dir_private}{The directory containing \code{file}. If NULL or
    missing, we use the user's home directory \code{"~"}. Use \code{""}
    if the \code{file} argument includes the directory.}
}
\value{
Returns value from \code{\link[gmailR]{gmail}}.
}
\description{
Send a text to yourself, using google mail info from a private file
}
\details{
Uses gmail address and password information in the
\code{~/.gmail_private} file (or another file, specified with
\code{dir_private} and \code{file_private}. This should contain records
\code{gmail}, \code{password}, and \code{text}, and (optionally) \code{to}.

An email will be sent from the address defined in the private file as \code{gmail}.
It will be sent to the email address defined as \code{text}. (If AT\&T is your
mobile phone company, this would be an address like \code{1234567890@txt.att.net}.)
}
\examples{
\dontrun{
send_text("R message", "R process done")
}
}
\seealso{
\code{\link{send_gmail}}
}
\keyword{utilities}

