% Generated by roxygen2 (4.0.1): do not edit by hand
\name{send_text}
\alias{send_text}
\title{Send a text to yourself}
\usage{
send_text(subject = "", body = "", body_sep = "\\n",
  file_private = "gmail_private.txt", dir_private = NULL)
}
\arguments{
\item{subject}{Character string with subject for the email}

\item{body}{Character string or vector of character strings, with
the body of the email; if it's a vector, it's combined with
\code{collapse=body_sep} (default = \code{"\n"}).}

\item{body_sep}{Character string for combining elements in the input \code{body}.}

\item{file_private}{Character string with file name containing the private information.}

\item{dir_private}{The directory containing \code{file}. If NULL or
    missing, we use the user's home directory \code{"~"}. Use \code{""}
    if the \code{file} argument includes the directory.}
}
\value{
Returns value from \code{\link[gmailr]{gmail}}.
}
\description{
Send a text to yourself, using google mail info from a private file
}
\details{
The private file should contain your gmail address, password, and
an email address corresponding to texting to your cell phone.
}
\examples{
\dontrun{
send_text("R message", "R process done")
}
}
\seealso{
\code{\link{send_gmail}}
}
\keyword{utilities}

